add angularjs-captcha npm

+++ ecommerce/src/main/java/com/shopping4th/ecommerce/service/ICategoryService.java
@@ -8,13 +8,13 @@ public interface ICategoryService {
 
 	public List<Category> findAll();
 	
-	public Category findById(int id);
+	public Category findById(Long id);
 	
 	public void save(Category category);
 	
-	public void deletedById(int id);
+	public void deletedById(Long id);
 	
-	public boolean existsById(int id);
+	public boolean existsById(Long id);
 	
 	
 }
