+++ ecommerce/src/main/java/com/shopping4th/ecommerce/service/IProductService.java
@@ -2,18 +2,17 @@ package com.shopping4th.ecommerce.service;
 
 import java.util.List;
 
-import com.shopping4th.ecommerce.entity.Category;
 import com.shopping4th.ecommerce.entity.Product;
 
 public interface IProductService {
 
 	public List<Product> findAll();
 	
-	public Product findById(int id);
+	public Product findById(Long id);
 	
 	public void save(Product product);
 	
-	public void deletedById(int id);
+	public void deletedById(Long id);
 	
-	public boolean existsById(int id);
+	public boolean existsById(Long id);
 }
